+++ test/ShapeCrawler.Tests.Unit/TableTests.cs
@@ -15,7 +15,6 @@ namespace ShapeCrawler.Tests.Unit;
 
 [SuppressMessage("ReSharper", "SuggestVarOrType_SimpleTypes")]
 [SuppressMessage("ReSharper", "SuggestVarOrType_BuiltInTypes")]
-[SuppressMessage("Usage", "xUnit1013:Public method should be marked as test")]
 public class TableTests : SCTest
 {
     [Xunit.Theory]